Move location of teacher
