+++ parlai/tasks/wrapper/README.md
@@ -1,0 +1,2 @@
+Teachers that wrap around other teachers.
+- `LabelToTextTeacher`: teacher that will shift the label of a message into the message's text field for the specified task
