+++ parlai/tasks/wrapper/agents.py
@@ -4,7 +4,7 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 """
-Generalized and miscellaneous teachers.
+Teachers that wrap around other ones.
 """
 
 import copy
